Separate out HCA packing code
